Rename ICES controller
